sh_nptl: Fix pthread_barrier_wait epilog to properly restore r9
In the pthread_barrier_wait function's epilog implementation do not
use the macro ret (that is defined as rts; nop) to correctly execute
in the delay slot the instruction  "mov.l    @r15+, r9" that will
restore the r9 register to the previous value saved on the stack and
as side effect will restore the stack pointer register back to the correct value.

Same fix is applied to __lll_mutex_lock_wait, even if in the delay slot
we have actually a nop, so using the macro would be safe here.

Signed-off-by: Christian Bruel <christian.bruel@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>