enable nanosecond stat support for everyone
Fill out the stat structure so that the nanosecond resolution support is
always available.  There is a small code size increase for a few ports
(three additional assignments in xstatconv), but otherwise everything
should remain the same.

While we're here, punt __old_kernel_stat from the few headers that still
define it as it is unused in uClibc and causes compile errors after these
nanosecond changes.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>