daemon: fix up INTERNAL_SYSCALL() usage
Make sure we declare the error properly in case a port uses it, and fix the
invocation of exit().  Since clone() will be returning a pid, assume that
the value will not have the MSB set (negative) to simplify the error test a
bit more.  If gcc supports it, force this function to always be heavily
optimized in a bid to avoid stack usage as much as possible.

Signed-off-by: James Coleman <james.coleman@ubicom.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>