sh/user.h: clean up conflicts from kernel user.h
This patch makes sufficient modifications to the SH sys/user.h to be
able to build and run gdbserver with Linux kernel 2.6.30.

The problem is that both the kernel and library headers define the same
symbols, thus making it impossible in include both headers. The patch
undefines the kernel symbols in favour of the library symbols.

Signed-off-by: Andrew Stubbs <ams@codesourcery.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>