LT pthread_atfork: unhide
Trying to compile perl with uClibc new linuxthreads on arm.
I run into the problem that pthread_atfork is not available.
The problem was that it was synced with the glibc version, which has
compat ifdefs we do not need in uClibc. The inital checked in version
is right and works. So either revert the commit 2 years ago
or patch it with the patch attached I made for openembedded.

Signed-off-by: Henning Heinold <heinold@inf.fu-berlin.de>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>