Fix ctime() standard compliance bug
fixes issue2209:
ctime() was updated in 0.9.31 to call localtime_r() instead of
localtime() to
avoid using a static buffer. Unfortunately, this change replaces the
static
buffer (which is zeroed out on initialization) with an uninitialized
local
buffer.

In the common case, this has no effect. However, with a sufficiently
large
time_t value, the value returned differs from that returned by
asctime(localtime(t)), and thus violates the ANSI/ISO standard.

An example input is (on a 64-bit machine):
time_t t = 0x7ffffffffff6c600;

Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>