buildsys: headers target should not depend on sysnum.h
The generation of the sysnum.h has to be a prerequisite of
'pregen' target instead of headers, because it needs the compiler
already available.
The so called C bootstrap headers are required to build the compiler
by invoking 'make headers', so the 'headers' target cannot relies itself
on the compiler to be invoked.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>