ldso: bfin: add alias from __start to _start
On Blackfin systems, we have underscore prefixes in our ABI.  So instead
of _start, we use __start.  However, the makefile explicitly sets _start
as the entry point to support toolchains that implicitly use other symbols
as the default entry.  It's easy enough to add a hidden alias in the
Blackfin code for free to support this.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>