pread/pwrite: backport fix
pread/pwrite syscalls has been renamed to pread64/pwrite in 2.6 kernel.

There was a fallback function using lseek for kernels who did not have
this syscall (pre 2.1.60). This is broken in many ways.

uclibc have been using the broken fallback due to they forgot to rename
pread syscall.

This got detected with git-1.7.11 which introduced threaded index-pack
which broke in similar ways a windows (msys).

This issue in uclibc have been reported upstream and fixed in git master
so this patch does not need to be upstreamed. It might be an idea to
backport it properly for 0.9.33 branch though.

Signed-off-by: Natanael Copa <natanael.copa@gmail.com>
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>